+++ Source/Project64/User Interface/WTL Controls/numberctrl.h
@@ -45,7 +45,7 @@ protected:
 		GetWindowText(WindowText,sizeof(WindowText));
 
 		bool bPaste=true;
-		unsigned int Len = strlen(WindowText);
+		size_t Len = strlen(WindowText);
 		char head   = Len > 0 ? WindowText[0] : 0;
 		char second = Len > 1 ? WindowText[1] : 0;
 
@@ -180,7 +180,7 @@ protected:
 		//CString text;
 		char WindowText[200];
 		GetWindowText(WindowText,sizeof(WindowText));
-		int Len = strlen(WindowText);
+		size_t Len = strlen(WindowText);
 
 		char head   = Len > 0 ? WindowText[0] : 0;
 		char second = Len > 1 ? WindowText[1] : 0;
@@ -294,14 +294,14 @@ public:
 			return atoi(text);
 		} 
 
-		int Finish = strlen(text);
+		size_t Finish = strlen(text);
 		char second = Finish > 1 ? text[1] : 0;
 		int Start = (second == 'x' || second == 'X') ? 2 : 0;
 
 		if (Finish > (8 + Start)) { Finish = (8 + Start); }
 		
 		DWORD Value = 0;
-		for (int i = Start; i < Finish; i++)
+		for (size_t i = Start; i < Finish; i++)
 		{
 			Value = (Value << 4);
 			switch( text[i] ) {
